Add files via upload
This is the first project form JSW plant. In this project what is main aim to achive, inside the plant some place are there where they indicate this is warning place. If any person will reach the particulare area we need take the time also when he out of the area that time also we need to track the time. For this we have used SSD mode to detect the object and cv2. Line used for tracking the time.